# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM API",
        "description": "Микро-CRM для управления заказами, остатками и движением товаров",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Основной сервер"
        }
    ],
    "paths": {
        "/api/orders": {
            "get": {
                "summary": "Получить список заказов",
                "operationId": "2cd7e03021e72d9d799c12dfe179c865",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтр по статусу",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "summary": "Создать заказ",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer": {
                                        "type": "string"
                                    },
                                    "warehouse_id": {
                                        "type": "integer"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer"
                                                },
                                                "count": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Обновить заказ (только клиент и товары)",
                "operationId": "6a7029f5510ca08e28a2d58a5fe19439",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заказа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Данные для обновления заказа",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer",
                                    "items"
                                ],
                                "properties": {
                                    "customer": {
                                        "type": "string",
                                        "example": "Иван Иванов"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "count": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешное обновление заказа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "customer": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "product_id": {
                                                        "type": "integer"
                                                    },
                                                    "count": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса"
                    },
                    "404": {
                        "description": "Заказ не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/orders/{id}/complete": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Завершить заказ (списание остатков)",
                "operationId": "6825d825bb4f63059e1a7363ce848760",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заказа для завершения",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заказ успешно завершен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "completed"
                                        },
                                        "completed_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка (например, недостаточно остатков или заказ уже не активен)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not enough stock for product ID 1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не найден"
                    }
                }
            }
        },
        "/api/orders/{id}/cancel": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Отменить заказ (возврат остатков на склад)",
                "operationId": "0181d49c16608e41e0f43658716300de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заказа для отмены",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заказ успешно отменён и остатки возвращены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "canceled"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка — можно отменить только завершённый заказ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Only completed orders can be canceled"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не найден"
                    }
                }
            }
        },
        "/api/orders/{id}/resume": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Возобновить отменённый заказ (снова списать остатки)",
                "operationId": "44b013b7b2f54d84d7048ac5047ae598",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID отменённого заказа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заказ успешно возобновлён и остатки списаны",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "completed"
                                        },
                                        "completed_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка: недостаточно остатков или заказ не отменён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not enough stock for product ID 3"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не найден"
                    }
                }
            }
        },
        "/api/stock-movements": {
            "get": {
                "tags": [
                    "Stock Movements"
                ],
                "summary": "Просмотр истории движений остатков",
                "operationId": "fde37fff909c0ea8acd51d19a5ce6c28",
                "parameters": [
                    {
                        "name": "warehouse_id",
                        "in": "query",
                        "description": "ID склада",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "ID товара",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Начальная дата (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-07-01"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Конечная дата (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-07-22"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество записей на страницу",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ: список движений остатков",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "product_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "warehouse_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "quantity_change": {
                                                        "type": "integer",
                                                        "example": -3
                                                    },
                                                    "reason": {
                                                        "type": "string",
                                                        "example": "order_completed"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-07-22T15:30:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 100
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Orders",
            "description": "Orders"
        },
        {
            "name": "Stock Movements",
            "description": "Stock Movements"
        }
    ]
}